Add more debugging logs
