+++ package.json
@@ -14,6 +14,7 @@
   "devDependencies": {
     "babel-eslint": "^8.0.1",
     "chalk": "^2.4.2",
+    "debug": "^4.1.1",
     "eslint": "^4.9.0",
     "eslint-config-formidable": "^3.0.0",
     "eslint-plugin-filenames": "^1.2.0",
